requirements.txt
braintree==4.10.0
channels==3.0.3
DateTimeRange==1.0.0
django-crispy-forms==1.11.2
Django==3.1
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-braces==1.14.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-npm==1.0.0
django-paypal==1.1.1
django-pipeline==2.0.6
django-rest-auth==0.9.5
django-import-export==2.5.0
django-phonenumber-field==5.0.0
django-rest-auth[with_social]
django-rest-authemail==2.0.6
django-s3-storage==0.13.4
django-select2==7.7.1
django-social-share==2.2.1
django-solo==1.1.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
fpdf==1.7.2
# fitz==0.0.1.dev2
jsmin==2.2.2
paypal-checkout-serversdk==1.0.1
pdf2image==1.16.0
phonenumbers==8.12.21
Pillow==8.2.0
psycopg2-binary==2.8.6
PyPDF2==1.26.0
PyMuPDF==1.18.17
reportlab==3.5.68
remote-pdb==2.1.0
setuptools==56.0.0
# signals==0.0.2
# utils==1.0.1
validators==0.18.2
whitenoise==5.2.0
zappa==0.52.0
zappa-django-utils==0.4.1
reportlab==3.5.68
django-autoslug==1.9.8
django-ckeditor==6.1.0
mammoth==1.4.17
PyDocX==0.9.10
